\name{logKcalc-package}
\alias{logKcalc-package}
\docType{package}
\title{Package Overview and Datasets}

\description{

\pkg{logKcalc} updates thermodynamic data files using equilibrium constants of reactions calculated from the OBIGT database in \pkg{CHNOSZ}.
The main function is \code{\link{logKcalc}}, which requires an existing GWB data file as input and creates a new output file with the calculated equilibrium constants.
\code{\link{logKcomp}} makes plots to compare equilibrium constants in the two files.

Some data files used for the examples and vignettes are available in the \code{extdata} directory of the package.

\samp{thermo_24elements.tdat}, with dataset format \code{jan19}, was generated from the GWB data file \samp{thermo.tdat} using GWB's TEdit program to remove elements other than Ag, Al, As, Au, C, Ca, Cl, Cu, F, Fe, H, K, Mg, Mn, N, Na, O, P, Pb, S, Si, Sn, U, and Zn.
The elements include those present in many of the GWB example scripts along with some others of particular interest for developing the OBIGT database (e.g. Au).
The original GWB data file \samp{thermo.tdat} is available in the \code{Gtdata} directory of the GWB installation and at \url{https://www.gwb.com/thermo.php}.

\samp{thermo_24swapped.tdat} is based on the previous file, with As(OH)3, Sn++, and UO2++ swapped for the basis species As(OH)4-, Sn++++, and U++++, respectively.
The result is a data file with basis species that are all in the default Unitherm and OBIGT databases.
This file is used by \code{\link{addOBIGT}} to regress thermodynamic parameters from logK values for species that are not in OBIGT.

\samp{thermo_24noredox.tdat} is based on the previous file, with basis species swapped for all redox species appearing in the dissociation reactions of other species.
The former redox species, which no longer appear in any reactions, were moved into the aqueous species block.
By writing all dissociation reactions in terms of basis species, the \logK values in this file can be compared with those in the file produced by K2GWB, where no redox species are present (see \code{\link{logKcomp}}).

\samp{ThermoGWB_15_6_2020.tdat}, with dataset format \code{oct94}, was created using K2GWB (Cleverley and Bastrakov, 2005) from the default Unitherm database in HCh version 4.4 (Shvarov, 2008).
This data file has the same basis species as \samp{thermo_24noredox.tdat}.
The suffix of the file output by K2GWB was changed from \code{dat} to \code{tdat}.
To match the names used in \samp{thermo_24noredox.tdat}, the output of K2GWB was manually modified to rename H2O(g), Oxygen(g), and H3AsO3(aq) to Steam, O2(g), and As(OH)3, respectively.

\samp{thermo_12elements.tdat} is an even smaller extract of \samp{thermo.tdat} with the elements Ag, C, Cl, Fe, H, K, N, Na, O, P, S, and Si.
This file is used for some examples and tests.

The directory \code{extdata/tests} contains files generated by \code{\link{logKcalc}} that are used for the package tests in \code{tests/testthat}.

}

\references{
Cleverley, J. S. and Bastrakov, E. N. (2005) K2GWB: Utility for generating thermodynamic data files for The Geochemist's Workbench® at 0-1000 °C and 1-5000 bar from UT2K and the UNITHERM database. \emph{Computers & Geosciences} \bold{31}, 756--767. \doi{10.1016/j.cageo.2005.01.007}

Shvarov, Yu. V. (2008) HCh: New potentialities for the thermodynamic simulation of geochemical systems offered by Windows. \emph{Geochemistry International} \bold{46}, 834--839. \doi{10.1134/S0016702908080089}
}

\keyword{package}
